clean up newregistrarsspec variables
